+++ src/scorer/feature/retoots_in_feed_scorer.ts
@@ -19,7 +19,7 @@ export default class RetootsInFeedScorer extends FeatureScorer {
         // add 1 if both reblog & toot are followed accounts
         const reblog = toot.reblog;
         let retootCount = reblog.account.isFollowed ? 1 : 0;
-        const nonAuthorRetoots = reblog.reblogsBy.filter((acct) => acct.webfingerURI != reblog.account.webfingerURI);
+        const nonAuthorRetoots = reblog.reblogsBy.filter((account) => account.webfingerURI != reblog.account.webfingerURI);
         retootCount += nonAuthorRetoots.length;
 
         // If retootsCount is 1 that's a normal retoot so we score it zero, otherwise return the square of retootCount