+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # NEXT RELEASE
 
+# v0.71.0
+* Add ability to filter by originating server
+* Don't sort `Toot` objects twice in `buildToots()`
+
 ### v0.70.5
 * Bump `brace-expansion` with `npm audit fix`
 
@@ -10,7 +14,7 @@
 * Uniquify `Account` objects by `url`, not by `id` (IDs are only unique across a single server)
 
 ### v0.70.2
-* Export `optionalSuffix()`
+* Export `optionalSuffix()` helper
 * JSDoc comments for `string_helpers.ts` and `Logger`
 
 ### v0.70.1